lib/device-map.py: add rpi3 32-bit support
rpi3 supports 32-bit mode and 64-bit mode kernels, but has a
dependency that the bootloader (typically u-boot) on the board
matches.  Therefore, upstream LAVA has a seprate device-type for
32-bit mode rpi3: bcm2837-rpi-3-b-32

Add support for generating 32-bit jobs for 32-bit ARM builds.

Note that currently (as of v4.15) only the bcm2835_defconfig is known
to be working with the upstream LAVA device type.

Signed-off-by: Kevin Hilman <khilman@baylibre.com>